pyproject.toml
[tool.poetry]
name = "fullgraph-alchemist"
version = "0.1.0"
description = "kekw"
authors = ["Dizzzma"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
flask_graphql = "^2.0.1"
graphene = "^2.1.8"
flask_sqlalchemy = "^2.4.4"
graphene_sqlalchemy = "^2.3.0"
smart_getenv = "^1.1.0"
pytest = "^5.4.3"
psycopg2-binary = "^2.8.5"
flask-script = "^2.0.6"
faker = "^4.1.1"
Jetkit-Flask = "^9.0.1"
factory-boy = "^2.12.0"
Flask-Migrate = "^2.5.3"
autoenv = "^1.0.0"
black = "^19.10b0"
pre-commit = "^2.6.0"
python-dotenv = "^0.14.0"
Flask-GraphQL-Auth = "^1.3.2"
flask_graphql_auth = "^1.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
